Update bot.js
Fixed 8 ball to accept proper questions and reject anything that isn't a question. This is based off the first word in the question which determines it as one. originally was a for loop with a nested if statement. I then discovered the includes function for arrays,  which means I can just put the question checker into an if and else statement.